+++ include/MMSP.grid.cpp
@@ -1,198 +1,7 @@
-// MMSP.grid.hpp
-// MMSP grid class definition and implementation
-// Questions/comments to gruberja@gmail.com (Jason Gruber)
+namespace MMSP{
 
-#ifndef MMSP_GRID
-#define MMSP_GRID
-#include<iostream>
-#include<cstdlib>
-#include<cstdarg>
-#include<sstream>
-#include<cmath>
-#include<limits>
-#include<cassert>
-#include <sys/statvfs.h>
-#ifndef RAW
-#include<zlib.h>
-#endif
-
-#include"MMSP.utility.hpp"
-#include"MMSP.scalar.hpp"
-#include"MMSP.vector.hpp"
-#include"MMSP.sparse.hpp"
-
-namespace MMSP
-{
-
-// declaration of grid class
-template <int dim, typename T> class grid;
-
-// grid utility functions
-template <int dim, typename T> int nodes(const grid<dim, T>& GRID)
-{
-	return nodes(GRID);
-}
-template <int dim, typename T> int fields(const grid<dim, T>& GRID)
-{
-	return fields(GRID);
-}
-template <int dim, typename T> int ghosts(const grid<dim, T>& GRID)
-{
-	return ghosts(GRID);
-}
-template <int dim, typename T> int g0(const grid<dim, T>& GRID, int i)
-{
-	return g0(GRID, i);
-}
-template <int dim, typename T> int g1(const grid<dim, T>& GRID, int i)
-{
-	return g1(GRID, i);
-}
-template <int dim, typename T> int b0(const grid<dim, T>& GRID, int i)
-{
-	return b0(GRID, i);
-}
-template <int dim, typename T> int b1(const grid<dim, T>& GRID, int i)
-{
-	return b1(GRID, i);
-}
-template <int dim, typename T> int& b0(grid<dim, T>& GRID, int i)
-{
-	return b0(GRID, i);
-}
-template <int dim, typename T> int& b1(grid<dim, T>& GRID, int i)
-{
-	return b1(GRID, i);
-}
-
-// grid utility functions (all directions)
-template <int dim, typename T> int x0(const grid<dim, T>& GRID, int i)
-{
-	return x0(GRID, i);
-}
-template <int dim, typename T> int x1(const grid<dim, T>& GRID, int i)
-{
-	return x1(GRID, i);
-}
-template <int dim, typename T> int xmin(const grid<dim, T>& GRID, int i)
-{
-	return xmin(GRID, i);
-}
-template <int dim, typename T> int xmax(const grid<dim, T>& GRID, int i)
-{
-	return xmax(GRID, i);
-}
-template <int dim, typename T> int xlength(const grid<dim, T>& GRID, int i)
-{
-	return xlength(GRID, i);
-}
-template <int dim, typename T> double dx(const grid<dim, T>& GRID, int i)
-{
-	return dx(GRID, i);
-}
-template <int dim, typename T> double& dx(grid<dim, T>& GRID, int i)
-{
-	return dx(GRID, i);
-}
-
-// grid utility functions (x direction)
-template <int dim, typename T> int x0(const grid<dim, T>& GRID)
-{
-	return x0(GRID);
-}
-template <int dim, typename T> int x1(const grid<dim, T>& GRID)
-{
-	return x1(GRID);
-}
-template <int dim, typename T> int xmin(const grid<dim, T>& GRID)
-{
-	return xmin(GRID);
-}
-template <int dim, typename T> int xmax(const grid<dim, T>& GRID)
-{
-	return xmax(GRID);
-}
-template <int dim, typename T> int xlength(const grid<dim, T>& GRID)
-{
-	return xlength(GRID);
-}
-template <int dim, typename T> double dx(const grid<dim, T>& GRID)
-{
-	return dx(GRID);
-}
-template <int dim, typename T> double& dx(grid<dim, T>& GRID)
-{
-	return dx(GRID);
-}
-
-// grid utility functions (y direction)
-template <int dim, typename T> int y0(const grid<dim, T>& GRID)
-{
-	return y0(GRID);
-}
-template <int dim, typename T> int y1(const grid<dim, T>& GRID)
-{
-	return y1(GRID);
-}
-template <int dim, typename T> int ymin(const grid<dim, T>& GRID)
-{
-	return ymin(GRID);
-}
-template <int dim, typename T> int ymax(const grid<dim, T>& GRID)
-{
-	return ymax(GRID);
-}
-template <int dim, typename T> int ylength(const grid<dim, T>& GRID)
-{
-	return ylength(GRID);
-}
-template <int dim, typename T> double dy(const grid<dim, T>& GRID)
-{
-	return dy(GRID);
-}
-template <int dim, typename T> double& dy(grid<dim, T>& GRID)
-{
-	return dy(GRID);
-}
-
-// grid utility functions (z direction)
-template <int dim, typename T> int z0(const grid<dim, T>& GRID)
-{
-	return z0(GRID);
-}
-template <int dim, typename T> int z1(const grid<dim, T>& GRID)
-{
-	return z1(GRID);
-}
-template <int dim, typename T> int zmin(const grid<dim, T>& GRID)
-{
-	return zmin(GRID);
-}
-template <int dim, typename T> int zmax(const grid<dim, T>& GRID)
-{
-	return zmax(GRID);
-}
-template <int dim, typename T> int zlength(const grid<dim, T>& GRID)
-{
-	return zlength(GRID);
-}
-template <int dim, typename T> double dz(const grid<dim, T>& GRID)
-{
-	return dz(GRID);
-}
-template <int dim, typename T> double& dz(grid<dim, T>& GRID)
-{
-	return dz(GRID);
-}
-
-
-// instantiation of grid class
-template <int dim, typename T>
-class grid
-{
-public:
-	// constructors
-	grid(int FIELDS, int min[dim], int max[dim], int GHOSTS = 1, bool SINGLE = false)
+template<int dim,typename T>
+grid<dim,T>::grid(int FIELDS, int min[dim], int max[dim], int GHOSTS, bool SINGLE)
 	{
 		// set number of fields
 		fields = FIELDS;
@@ -214,7 +23,8 @@ public:
 		setup(SINGLE);
 	}
 
-	grid(int FIELDS, ...)
+template<int dim,typename T>
+grid<dim,T>::grid(int FIELDS, ...)
 	{
 		// set number of fields
 		fields = FIELDS;
@@ -239,8 +49,8 @@ public:
 		// setup grid properties
 		setup();
 	}
-
-	grid(const grid& GRID)
+template<int dim,typename T>
+grid<dim,T>::grid(const grid& GRID)
 	{
 		// set number of fields
 		fields = MMSP::fields(GRID);
@@ -269,8 +79,8 @@ public:
 		}
 	}
 
-	template <typename U>
-	grid(const grid<dim, U>& GRID, int FIELDS)
+template <int dim, typename T> template<typename U>
+grid<dim,T>::grid(const grid<dim,U>& GRID, int FIELDS)
 	{
 		// set number of fields
 		fields = FIELDS;
@@ -298,8 +108,8 @@ public:
 			dx[i] = MMSP::dx(GRID, i);
 		}
 	}
-
-	grid(const char* filename, int GHOSTS = 1)
+template<int dim,typename T>
+grid<dim,T>::grid(const char* filename, int GHOSTS)
 	{
 		// initialize data
 		data=NULL;
@@ -307,9 +117,8 @@ public:
 		// read data from file
 		input(filename, GHOSTS);
 	}
-
-	void setup(bool SINGLE = false)
-	{
+template<int dim, typename T>
+void grid<dim,T>::setup(bool SINGLE){
 		// setup default grid parameters
 		for (int i=0; i<dim; i++) {
 			x0[i] = g0[i];
@@ -493,9 +302,8 @@ public:
 		MPI::COMM_WORLD.Barrier();
 		#endif
 	}
-
-	// destructor
-	~grid()
+template<int dim, typename T>
+grid<dim,T>::~grid()
 	{
 		if (data!=NULL) {
 			delete [] data;
@@ -503,47 +311,40 @@ public:
 		}
 	}
 
+// assignment operators
 
-	// assignment operators
-	template <typename U> grid& operator=(const U& value)
+template<int dim, typename T> template <typename U> grid<dim,T>& grid<dim,T>::operator=(const U& value)
 	{
 		for (int i=0; i<cells; i++)
 			data[i] = static_cast<T>(value);
-		return *this;
 	}
 
-	template <typename U> grid& operator=(const grid<dim, U>& GRID)
+template<int dim, typename T> template <typename U> grid<dim,T>& grid<dim,T>::operator=(const grid<dim, U>& GRID)
 	{
-		if (this == &GRID)
-			return *this;
 		for (int i=0; i<cells; i++)
 			data[i] = static_cast<T>(GRID.data[i]);
-		return *this;
 	}
 
-	template <typename U> grid& operator+=(const grid<dim, U>& GRID)
+template<int dim, typename T> template <typename U> grid<dim,T>& grid<dim,T>::operator+=(const grid<dim, U>& GRID)
 	{
 		for (int i=0; i<cells; i++)
 			data[i] += static_cast<T>(GRID.data[i]);
-		return *this;
 	}
 
-	template <typename U> grid& operator-=(const grid<dim, U>& GRID)
+template<int dim, typename T> template <typename U> grid<dim,T>& grid<dim,T>::operator-=(const grid<dim, U>& GRID)
 	{
 		for (int i=0; i<cells; i++)
 			data[i] -= static_cast<T>(GRID.data[i]);
-		return *this;
 	}
 
-
 	// subscript operators
-	target < dim - 1, 0, T > operator [](int x) const
+template<int dim, typename T> target < dim - 1, 0, T > grid<dim,T>::operator [](int x) const
 	{
 		check_boundary(x, x0[0], x1[0], b0[0], b1[0]);
 		return target < dim - 1, 0, T > (data + (x - s0[0]) * sx[0], s0, sx, x0, x1, b0, b1);
 	}
 
-	T& operator()(MMSP::vector<int> x) const
+template<int dim, typename T> T& grid<dim,T>::operator()(MMSP::vector<int> x) const
 	{
 		T* p = data;
 		for (int i=0; i<dim; i++) {
@@ -553,7 +354,7 @@ public:
 		return *p;
 	}
 
-	T& operator()(int i) const
+template<int dim, typename T> T& grid<dim,T>::operator()(int i) const
 	{
 		#ifdef MPI_VERSION
 		int x[dim];
@@ -570,9 +371,7 @@ public:
 		return data[i];
 	}
 
-
-	// position utility function
-	MMSP::vector<int> position(int i) const
+template<int dim, typename T> MMSP::vector<int> grid<dim,T>::position(int i) const
 	{
 		MMSP::vector<int> x(dim);
 		for (int j=0; j<dim; j++) {
@@ -583,9 +382,8 @@ public:
 		return x;
 	}
 
-
 	// parallelization
-	void ghostswap()
+template<int dim, typename T> void grid<dim,T>::ghostswap()
 	{
 		#ifdef MPI_VERSION
 		MPI::COMM_WORLD.Barrier();
@@ -683,8 +481,7 @@ public:
 		MPI::COMM_WORLD.Barrier();
 		#endif
 	}
-
-	void ghostswap(const int sublattice)   // ghostswap for Monte Carlo communiation reduction.
+template<int dim, typename T> void grid<dim,T>::ghostswap(const int sublattice)   // ghostswap for Monte Carlo communiation reduction.
 	{
 		#ifdef MPI_VERSION
 		MPI::COMM_WORLD.Barrier();
@@ -1217,12 +1014,11 @@ public:
 		MPI::COMM_WORLD.Barrier();
 		#endif
 	}
-
-	unsigned long buffer_size_save(const int min[dim], const int max[dim]) const
+template<int dim,typename T> unsigned long grid<dim,T>::buffer_size_save(const int min[dim], const int max[dim]) const
 	{
 		return buffer_size_save(data, 0, min, max);
 	}
-	unsigned long buffer_size_save(T* p, int i, const int min[dim], const int max[dim]) const
+template<int dim, typename T> unsigned long grid<dim,T>::buffer_size_save(T* p, int i, const int min[dim], const int max[dim]) const
 	{
 		unsigned long size = 0;
 		if (i == dim - 1)
@@ -1235,12 +1031,12 @@ public:
 		return size;
 	}
 
-	unsigned long to_buffer_save(char* buffer, const int min[dim], const int max[dim]) const
+template<int dim, typename T> unsigned long grid<dim,T>::to_buffer_save(char* buffer, const int min[dim], const int max[dim]) const
 	{
 		return to_buffer_save(buffer, data, 0, min, max);
 	}
 
-	unsigned long to_buffer_save(char* buffer, T* p, int i, const int min[dim], const int max[dim]) const
+template<int dim, typename T> unsigned long grid<dim,T>::to_buffer_save(char* buffer, T* p, int i, const int min[dim], const int max[dim]) const
 	{
 		unsigned long size = 0;
 		if (i == dim - 1)
@@ -1252,12 +1048,12 @@ public:
 		return size;
 	}
 
-	unsigned long from_buffer_save(char* buffer, const int min[dim], const int max[dim])
+template<int dim, typename T> unsigned long grid<dim,T>::from_buffer_save(char* buffer, const int min[dim], const int max[dim])
 	{
 		return from_buffer_save(buffer, data, 0, min, max);
 	}
 
-	unsigned long from_buffer_save(char* buffer, T* p, int i, const int min[dim], const int max[dim])
+template<int dim, typename T> unsigned long grid<dim,T>::from_buffer_save(char* buffer, T* p, int i, const int min[dim], const int max[dim])
 	{
 		unsigned long size = 0;
 		if (i == dim - 1) {
@@ -1270,17 +1066,7 @@ public:
 		return size;
 	}
 	// buffer I/O
-	unsigned long buffer_size() const
-	{
-		return buffer_size(x0, x1);
-	}
-
-	unsigned long buffer_size(const int min[dim], const int max[dim]) const
-	{
-		return buffer_size(data, 0, min, max);
-	}
-
-	unsigned long buffer_size(T* p, int i, const int min[dim], const int max[dim]) const
+template<int dim, typename T> unsigned long grid<dim,T>::buffer_size(T* p, int i, const int min[dim], const int max[dim]) const
 	{
 		unsigned long size = 0;
 		if (i == dim - 1)
@@ -1292,17 +1078,29 @@ public:
 		return size;
 	}
 
-	unsigned long to_buffer(char* buffer) const
+template<int dim, typename T> unsigned long grid<dim,T>::buffer_size() const
+	{
+		return buffer_size(x0, x1);
+	}
+
+template<int dim, typename T> unsigned long grid<dim,T>::buffer_size(const int min[dim], const int max[dim]) const
+	{
+		return buffer_size(data, 0, min, max);
+	}
+
+
+
+template<int dim, typename T> unsigned long grid<dim,T>::to_buffer(char* buffer) const
 	{
 		return to_buffer(buffer, x0, x1);
 	}
 
-	unsigned long to_buffer(char* buffer, const int min[dim], const int max[dim]) const
+template<int dim, typename T> unsigned long grid<dim,T>::to_buffer(char* buffer, const int min[dim], const int max[dim]) const
 	{
 		return to_buffer(buffer, data, 0, min, max);
 	}
 
-	unsigned long to_buffer(char* buffer, T* p, int i, const int min[dim], const int max[dim]) const
+template<int dim, typename T> unsigned long grid<dim,T>::to_buffer(char* buffer, T* p, int i, const int min[dim], const int max[dim]) const
 	{
 		unsigned long size = 0;
 		if (i == dim - 1)
@@ -1314,17 +1112,17 @@ public:
 		return size;
 	}
 
-	unsigned long from_buffer(char* buffer)
+template<int dim, typename T> unsigned long grid<dim,T>::from_buffer(char* buffer)
 	{
 		return from_buffer(buffer, x0, x1);
 	}
 
-	unsigned long from_buffer(char* buffer, const int min[dim], const int max[dim])
+template<int dim, typename T> unsigned long grid<dim,T>::from_buffer(char* buffer, const int min[dim], const int max[dim])
 	{
 		return from_buffer(buffer, data, 0, min, max);
 	}
 
-	unsigned long from_buffer(char* buffer, T* p, int i, const int min[dim], const int max[dim])
+template<int dim, typename T> unsigned long grid<dim,T>::from_buffer(char* buffer, T* p, int i, const int min[dim], const int max[dim])
 	{
 		unsigned long size = 0;
 		if (i == dim - 1) {
@@ -1336,16 +1134,15 @@ public:
 		}
 		return size;
 	}
-
 	// file I/O
-	void input(const char* filename, int GHOSTS = 1, int SINGLE = false)
+template<int dim, typename T> void grid<dim,T>::input(const char* filename, int GHOSTS, int SINGLE)
 	{
 		// file open error check
 		std::ifstream input(filename);
 		if (!input) {
 			std::cerr << "File input error: could not open ";
 			std::cerr << filename << "." << std::endl;
-			MMSP::Abort(-1);
+			exit(-1);
 		}
 
 		// grid data type error check
@@ -1353,7 +1150,7 @@ public:
 		getline(input, type, '\n');
 		if (type != name(*this)) {
 			std::cerr << "File read error: wrong data type (" << type << ")." << std::endl;
-			MMSP::Abort(-2);
+			exit(-2);
 		}
 
 		// dimension error check
@@ -1361,7 +1158,7 @@ public:
 		input >> dimen;
 		if (dimen != dim) {
 			std::cerr << "File read error: wrong dimension (" << dimen << ")." << std::endl;
-			MMSP::Abort(-3);
+			exit(-3);
 		}
 
 		// read number of fields
@@ -1399,7 +1196,7 @@ public:
 		input.close();
 	}
 
-	void read(std::ifstream& file, int GHOSTS = 1)
+template<int dim, typename T> void grid<dim,T>::read(std::ifstream& file, int GHOSTS)
 	{
 		/*
 			Reads each block of the input file and constructs a
@@ -1460,7 +1257,7 @@ public:
 				if (size_in_mem!=size_on_disk) {
 					#ifdef RAW
 					std::cerr<<"Unable to uncompress data: compiled without zlib."<<std::endl;
-					MMSP::Abort(1);
+					exit(1);
 					#else
 					// Uncompress data
 					char* raw = new char[size_in_mem];
@@ -1470,11 +1267,11 @@ public:
 						break;
 					case Z_MEM_ERROR:
 						std::cerr << "Uncompress: out of memory." << std::endl;
-						MMSP::Abort(1);    // quit.
+						exit(1);    // quit.
 						break;
 					case Z_BUF_ERROR:
 						std::cerr << "Uncompress: output buffer wasn't large enough." << std::endl;
-						MMSP::Abort(1);    // quit.
+						exit(1);    // quit.
 						break;
 					}
 					GRID.from_buffer(raw);
@@ -1508,8 +1305,7 @@ public:
 		MPI::COMM_WORLD.Barrier();
 		#endif
 	}
-
-	void output(const char* filename) const
+template<int dim, typename T> void grid<dim,T>::output(const char* filename) const
 	{
 		#ifndef MPI_VERSION
 		unsigned int np=1;
@@ -1518,7 +1314,7 @@ public:
 		if (!output) {
 			std::cerr << "File output error: could not open ";
 			std::cerr << filename << "." << std::endl;
-			MMSP::Abort(-1);
+			exit(-1);
 		}
 
 		std::stringstream outstr;
@@ -1575,22 +1371,19 @@ public:
 		if (blocksize<1048576) {
 			// Standard MPI-IO: every rank writes to disk
 			#ifdef BGQ
-			if (rank==0) std::cout<<"Bug: using MPI-IO, instead of BGQ IO!"<<std::endl;
+			if (rank==0) std::cout<<"Bug: using stock MPI-IO, instead of BGQ IO!"<<std::endl;
 			#endif
 			MPI_File output;
-			mpi_err = MPI_File_open(MPI_COMM_WORLD, fname, MPI_MODE_CREATE|MPI_MODE_EXCL|MPI_MODE_WRONLY, MPI_INFO_NULL, &output);
+			mpi_err = MPI_File_open(MPI_COMM_WORLD, fname, MPI_MODE_WRONLY|MPI_MODE_EXCL|MPI_MODE_CREATE, MPI_INFO_NULL, &output);
 			if (mpi_err != MPI_SUCCESS) {
 				if (rank==0)
 					MPI_File_delete(fname,MPI_INFO_NULL);
 				mpi_err = MPI_File_open(MPI_COMM_WORLD, fname, MPI_MODE_WRONLY|MPI_MODE_EXCL|MPI_MODE_CREATE, MPI_INFO_NULL, &output);
 				assert(mpi_err==MPI_SUCCESS);
 			}
-			if (mpi_err != MPI_SUCCESS) {
-				char error_string[256];
-				int length_of_error_string=256;
-				MPI_Error_string(mpi_err, error_string, &length_of_error_string);
-				fprintf(stderr, "Error opening output file %s on rank %3d: %s\n", fname, rank, error_string);
-				MMSP::Abort(-1);
+			if (!output) {
+				std::cerr << "File output error: could not open " << fname << "." << std::endl;
+				exit(-1);
 			}
 
 			// Generate MMSP header from rank 0
@@ -1663,11 +1456,11 @@ public:
 			#ifdef GRIDDEBUG
 			int error, write_errors=0;
 			MPI_Get_count(&status, MPI_INT, &error);
-			error++;
-			if (error!=1) std::cerr<<"  Error on Rank "<<rank<<": "<<MPI::Get_error_class(error-1)<<std::endl;
+			if (error!=0)
+				std::cerr<<"  Error on Rank "<<rank<<": "<<MPI::Get_error_class(error)<<std::endl;
 			MPI::COMM_WORLD.Allreduce(&error, &write_errors, 1, MPI_INT, MPI_SUM);
-			if (rank==0) std::cout<<"  Write finished on "<<write_errors<<'/'<<np<<" ranks."<<std::endl;
-			assert(write_errors==np);
+			if (rank==0) std::cout<<"  Write finished on "<<np-write_errors<<'/'<<np<<" ranks."<<std::endl;
+			assert(write_errors==0);
 			#endif
 			delete [] buffer;
 			buffer=NULL;
@@ -1901,20 +1694,23 @@ public:
 			if (rank==0) std::cout<<"  Opening "<<std::string(fname)<<" for output."<<std::endl;
 			#endif
 			MPI_File output;
-			MPI::COMM_WORLD.Barrier();
-			mpi_err = MPI_File_open(MPI_COMM_WORLD, fname, MPI_MODE_CREATE|MPI_MODE_EXCL|MPI_MODE_WRONLY, MPI_INFO_NULL, &output);
+			mpi_err = MPI_File_open(MPI_COMM_WORLD, fname, MPI_MODE_WRONLY|MPI_MODE_EXCL|MPI_MODE_CREATE, MPI_INFO_NULL, &output);
 			if (mpi_err != MPI_SUCCESS) {
 				if (rank==0)
 					MPI_File_delete(fname,MPI_INFO_NULL);
 				mpi_err = MPI_File_open(MPI_COMM_WORLD, fname, MPI_MODE_WRONLY|MPI_MODE_EXCL|MPI_MODE_CREATE, MPI_INFO_NULL, &output);
 				assert(mpi_err==MPI_SUCCESS);
 			}
+			if (!output) {
+				if (rank==0) std::cerr << "File output error: could not open " << fname << "." << std::endl;
+				if (rank==0) std::cerr << "                   If it already exists, delete it and try again." << std::endl;
+				exit(-1);
+			}
 			if (mpi_err != MPI_SUCCESS) {
 				char error_string[256];
 				int length_of_error_string=256;
 				MPI_Error_string(mpi_err, error_string, &length_of_error_string);
-				fprintf(stderr, "Error opening output file %s on rank %3d: %s\n", fname, rank, error_string);
-				MMSP::Abort(-1);
+				fprintf(stderr, "%3d: %s\n", rank, error_string);
 			}
 
 			// Write to disk
@@ -1963,8 +1759,7 @@ public:
 		}
 		#endif
 	}
-
-	unsigned long write_buffer(char*& buf) const
+template<int dim, typename T> unsigned long grid<dim,T>::write_buffer(char*& buf) const
 	{
 		// Find out how big the dataset is
 		unsigned long size_in_mem = this->buffer_size();
@@ -2036,11 +1831,11 @@ public:
 			break;
 		case Z_MEM_ERROR:
 			std::cerr << "Compress: out of memory." << std::endl;
-			MMSP::Abort(1);
+			exit(1);
 			break;
 		case Z_BUF_ERROR:
 			std::cerr << "Compress: output buffer wasn't large enough." << std::endl;
-			MMSP::Abort(1);
+			exit(1);
 			break;
 		}
 		assert(size_on_disk<=size_of_buffer); // Abort if data was lost in compression process. Duplicate of Z_BUF_ERROR check.
@@ -2056,188 +1851,10 @@ public:
 		// Return total size of the populated buffer
 		return header_size + static_cast<unsigned long>(sizeof(size_in_mem)) + static_cast<unsigned long>(sizeof(size_on_disk)) + size_on_disk;
 	}
-
 	// grid utility functions
-	friend int nodes(const grid& GRID)
-	{
-		return GRID.nodes;
-	}
-	friend int fields(const grid& GRID)
-	{
-		return GRID.fields;
-	}
-	friend int ghosts(const grid& GRID)
-	{
-		return GRID.ghosts;
-	}
-	friend int g0(const grid& GRID, int i)
-	{
-		return GRID.g0[i];
-	}
-	friend int g1(const grid& GRID, int i)
-	{
-		return GRID.g1[i];
-	}
-	friend int b0(const grid& GRID, int i)
-	{
-		return GRID.b0[i];
-	}
-	friend int b1(const grid& GRID, int i)
-	{
-		return GRID.b1[i];
-	}
-	friend int& b0(grid& GRID, int i)
-	{
-		return GRID.b0[i];
-	}
-	friend int& b1(grid& GRID, int i)
-	{
-		return GRID.b1[i];
-	}
-
-	// grid utility functions (all directions)
-	friend int x0(const grid& GRID, int i)
-	{
-		return GRID.x0[i];
-	}
-	friend int x1(const grid& GRID, int i)
-	{
-		return GRID.x1[i];
-	}
-	friend int xmin(const grid& GRID, int i)
-	{
-		return GRID.x0[i];
-	}
-	friend int xmax(const grid& GRID, int i)
-	{
-		return GRID.x1[i];
-	}
-	friend int xlength(const grid& GRID, int i)
-	{
-		return GRID.x1[i] - GRID.x0[i];
-	}
-	friend double dx(const grid& GRID, int i)
-	{
-		return GRID.dx[i];
-	}
-	friend double& dx(grid& GRID, int i)
-	{
-		return GRID.dx[i];
-	}
-	friend int N0(const grid& GRID, int i)
-	{
-		return GRID.n0[i];
-	}
-	friend int N1(const grid& GRID, int i)
-	{
-		return GRID.n1[i];
-	}
-	friend int P0(const grid& GRID, int i)
-	{
-		return GRID.p0[i];
-	}
-	friend int P1(const grid& GRID, int i)
-	{
-		return GRID.p1[i];
-	}
-	friend int sp(const grid& GRID, int i)
-	{
-		return GRID.sp[i];
-	}
-
-	// grid utility functions (x direction)
-	friend int x0(const grid& GRID)
-	{
-		return GRID.x0[0];
-	}
-	friend int x1(const grid& GRID)
-	{
-		return GRID.x1[0];
-	}
-	friend int xmin(const grid& GRID)
-	{
-		return GRID.x0[0];
-	}
-	friend int xmax(const grid& GRID)
-	{
-		return GRID.x1[0];
-	}
-	friend int xlength(const grid& GRID)
-	{
-		return GRID.x1[0] - GRID.x0[0];
-	}
-	friend double dx(const grid& GRID)
-	{
-		return GRID.dx[0];
-	}
-	friend double& dx(grid& GRID)
-	{
-		return GRID.dx[0];
-	}
-
-	// grid utility functions (y direction)
-	friend int y0(const grid& GRID)
-	{
-		return GRID.x0[1];
-	}
-	friend int y1(const grid& GRID)
-	{
-		return GRID.x1[1];
-	}
-	friend int ymin(const grid& GRID)
-	{
-		return GRID.x0[1];
-	}
-	friend int ymax(const grid& GRID)
-	{
-		return GRID.x1[1];
-	}
-	friend int ylength(const grid& GRID)
-	{
-		return GRID.x1[1] - GRID.x0[1];
-	}
-	friend double dy(const grid& GRID)
-	{
-		return GRID.dx[1];
-	}
-	friend double& dy(grid& GRID)
-	{
-		return GRID.dx[1];
-	}
-
-	// grid utility functions (z direction)
-	friend int z0(const grid& GRID)
-	{
-		return GRID.x0[2];
-	}
-	friend int z1(const grid& GRID)
-	{
-		return GRID.x1[2];
-	}
-	friend int zmin(const grid& GRID)
-	{
-		return GRID.x0[2];
-	}
-	friend int zmax(const grid& GRID)
-	{
-		return GRID.x1[2];
-	}
-	friend int zlength(const grid& GRID)
-	{
-		return GRID.x1[2] - GRID.x0[2];
-	}
-	friend double dz(const grid& GRID)
-	{
-		return GRID.dx[2];
-	}
-	friend double& dz(grid& GRID)
-	{
-		return GRID.dx[2];
-	}
-
-
+	
 	// utility functions
-	void swap(grid& GRID)
+template<int dim, typename T> void grid<dim,T>::swap(grid& GRID)
 	{
 		// swap grid data
 		T* DATA = data;
@@ -2316,8 +1933,7 @@ public:
 			GRID.n1[i] = N1;
 		}
 	}
-
-	void copy(const grid& GRID)
+template<int dim, typename T> void grid<dim,T>::copy(const grid& GRID)
 	{
 		// initialize data
 		if (data != NULL) {
@@ -2372,43 +1988,7 @@ public:
 			buffer=NULL;
 		}
 	}
-
-
-protected:
-	T* data;        // local grid data
-
-	int nodes;			// number of nodes (excluding ghosts)
-	int cells;			// number of nodes (including ghosts)
-	int fields;			// number of fields
-	int ghosts;			// ghost cell depth
-
-	int g0[dim];    // global lower coordinate limit (excluding ghosts)
-	int g1[dim];    // global upper coordinate limit (excluding ghosts)
-
-	int x0[dim];    // local lower coordinate limit (excluding ghosts)
-	int x1[dim];    // local upper coordinate limit (excluding ghosts)
-	int xx[dim];    // local cells in slice (excluding ghosts)
-
-	int s0[dim];    // local lower coordinate limit (including ghosts)
-	int s1[dim];    // local upper coordinate limit (including ghosts)
-	int sx[dim];    // local cells in slice (including ghosts)
-
-	int b0[dim];    // boundary condition at x0
-	int b1[dim];    // boundary condition at x1
-
-	double dx[dim]; // global cell spacing
-
-	int p0[dim];
-	int p1[dim];
-	int sp[dim];    // global processors in slice
-
-	int n0[dim];    // neighbor processor at x0
-	int n1[dim];    // neighbor processor at x1
-};
-
-
-// math functions
-template <int dim, typename T> T laplacian(const grid<dim, T>& GRID, const vector<int>& x)
+template <int dim, typename T> T laplacian(const MMSP::grid<dim, T>& GRID, const vector<int>& x)
 {
 	T laplacian = 0.0;
 	MMSP::vector<int> s = x;
@@ -2429,7 +2009,8 @@ template <int dim, typename T> T laplacian(const grid<dim, T>& GRID, const vecto
 
 template <int dim, typename T> vector<T> laplacian(const grid<dim, vector<T> >& GRID, const vector<int>& x)
 {
-	vector<T> laplacian(fields(GRID), 0.0);
+	int N = fields(GRID);
+	vector<T> laplacian(N, 0.0);
 	vector<int> s = x;
 
 	const vector<T>& y = GRID(x);
@@ -2442,14 +2023,15 @@ template <int dim, typename T> vector<T> laplacian(const grid<dim, vector<T> >& 
 		s[i] += 1;
 
 		double weight = 1.0 / (dx(GRID, i) * dx(GRID, i));
-		laplacian += weight * (yh - 2.0 * y + yl);
+		for (int j=0; j<N; j++)
+			laplacian[j] += weight * (yh[j] - 2.0 * y[j] + yl[j]);
 	}
 	return laplacian;
 }
 
 template<int dim, typename T> T laplacian(const grid<dim,vector<T> >& GRID, const vector<int>& x, const int field)
 {
-    T laplacian = 0.0;
+    double laplacian = 0.0;
     vector<int> s = x;
 
     const T& y = GRID(x)[field];
@@ -2511,7 +2093,7 @@ template <int dim, typename T> sparse<T> laplacian(const grid<dim, sparse<T> >& 
 	return laplacian(GRID, x);
 }
 
-template <int dim, typename T> vector<T> gradient(const grid<dim, T>& GRID, const vector<int>& x)
+template <int dim, typename T> MMSP::vector<T> gradient(const grid<dim, T>& GRID, const vector<int>& x)
 {
 	vector<T> gradient(dim);
 	vector<int> s = x;
@@ -2523,7 +2105,7 @@ template <int dim, typename T> vector<T> gradient(const grid<dim, T>& GRID, cons
 		const T& yl = GRID(s);
 		s[i] += 1;
 
-		double weight = 1.0 / (2.0 * dx(GRID, i));
+		double weight = 1.0 / (2.0 * dx(GRID, i) );
 		gradient[i] = weight * (yh - yl);
 	}
 	return gradient;
@@ -2534,27 +2116,25 @@ template <int dim, typename T> vector<T> grad(const grid<dim, T>& GRID, const ve
 	return gradient(GRID, x);
 }
 
-template <int dim, typename T> vector<T> gradient(const grid<dim, vector<T> >& GRID, const vector<int>& x, const int f)
+template <int dim, typename T> vector<T> divergence(const grid<dim, vector<T> >& GRID, const vector<int>& x)
 {
-	vector<T> gradient(dim);
+	vector<T> divergence(dim, 0.0);
 	vector<int> s = x;
+
+	int N = length(GRID(x));
 
 	for (int i=0; i<dim; i++) {
 		s[i] += 1;
-		const T& yh = GRID(s)[f];
+		const vector<T>& yh = GRID(s);
 		s[i] -= 2;
-		const T& yl = GRID(s)[f];
+		const vector<T>& yl = GRID(s);
 		s[i] += 1;
 
 		double weight = 1.0 / (2.0 * dx(GRID, i));
-		gradient[i] = weight * (yh - yl);
+		for (int j=0; j<N; j++)
+			divergence[j] += weight * (yh[j] - yl[j]);
 	}
-	return gradient;
-}
-
-template <int dim, typename T> vector<T> grad(const grid<dim, T>& GRID, const vector<int>& x, const int f)
-{
-	return gradient(GRID, x, f);
+	return divergence;
 }
 
 template <int dim, typename T> T divergence(const grid<dim, T>& GRID, const vector<int>& x)
@@ -2575,105 +2155,7 @@ template <int dim, typename T> T divergence(const grid<dim, T>& GRID, const vect
 	return divergence;
 }
 
-template <int dim, typename T> vector<T> divergence(const grid<dim, vector<T> >& GRID, const vector<int>& x)
-{
-	vector<T> divergence(fields(GRID), 0.0);
-	vector<int> s = x;
 
-	for (int i=0; i<dim; i++) {
-		s[i] += 1;
-		const vector<T>& yh = GRID(s);
-		s[i] -= 2;
-		const vector<T>& yl = GRID(s);
-		s[i] += 1;
-
-		double weight = 1.0 / (2.0 * dx(GRID, i));
-		divergence += weight * (yh - yl);
-	}
-	return divergence;
-}
-
-template <int dim, typename T> T div(const grid<dim, T>& GRID, const vector<int>& x)
-{
-	return divergence(GRID, x);
-}
-
-template <int dim, typename T> vector<T> div(const grid<dim, vector<T> >& GRID, const vector<int>& x)
-{
-	return divergence(GRID, x);
-}
-
-// position utility function
-template <int dim, typename T> MMSP::vector<int> position(const grid<dim, T>& GRID, int index)
-{
-	return GRID.position(index);
-}
-
-// parallelization
-template <int dim, typename T> void ghostswap(grid<dim, T>& GRID)
-{
-	GRID.ghostswap();
-}
-
-template <int dim, typename T> void ghostswap(grid<dim, T>& GRID, const int sublattice)
-{
-	GRID.ghostswap(sublattice);
-}
-// buffer I/O functions
-template <int dim, typename T> unsigned long buffer_size(const grid<dim, T>& GRID)
-{
-	return GRID.buffer_size();
-}
-template <int dim, typename T> unsigned long to_buffer(const grid<dim, T>& GRID, char* buffer)
-{
-	return GRID.to_buffer(buffer);
-}
-template <int dim, typename T> unsigned long from_buffer(grid<dim, T>& GRID, const char* buffer)
-{
-	return GRID.from_buffer(buffer);
-}
-
-// file I/O functions
-template <int dim, typename T> void read(grid<dim, T>& GRID, std::ifstream& file)
-{
-	GRID.read(file);
-}
-template <int dim, typename T> void write(const grid<dim, T>& GRID, std::ifstream& file)
-{
-	GRID.write(file);
-}
-template <int dim, typename T> void input(grid<dim, T>& GRID, const char* filename, int GHOSTS = 1, int SINGLE = false)
-{
-	GRID.input(filename, GHOSTS, SINGLE);
-}
-template <int dim, typename T> void output(const grid<dim, T>& GRID, const char* filename)
-{
-	GRID.output(filename);
-}
-template <int dim, typename T> unsigned long write_buffer(const grid<dim, T>& GRID, char*& buf)
-{
-	return GRID.write_buffer(buf);
-}
-
-// utility functions
-template <int dim, typename T> int length(const grid<dim, T>& GRID)
-{
-	return nodes(GRID);
-}
-template <int dim, typename T> void resize(int n, grid<dim, T>& GRID) {}
-template <int dim, typename T> void swap(grid<dim, T>& GRID1, grid<dim, T>& GRID2)
-{
-	GRID1.swap(GRID2);
-}
-template <int dim, typename T> void copy(grid<dim, T>& GRID1, grid<dim, T>& GRID2)
-{
-	GRID2.copy(GRID1);
-}
-template <int dim, typename T> std::string name(const grid<dim, T>& GRID)
-{
-	return std::string("grid:") + name(T());
-}
 
 } // namespace MMSP
 
-#endif
