+++ include/MMSP.hpp
@@ -2,11 +2,17 @@
 // All encompassing MMSP header file
 // Questions/comments to gruberja@gmail.com (Jason Gruber)
 
-#include"MMSP.utility.hpp"
+#include"MMSP.utility.h"
+#include"MMSP.utility.cpp"
 
-#include"MMSP.scalar.hpp"
-#include"MMSP.vector.hpp"
-#include"MMSP.sparse.hpp"
+#include"MMSP.scalar.h"
 
-#include"MMSP.grid.hpp"
+#include"MMSP.vector.h"
+#include"MMSP.vector.cpp"
+
+#include"MMSP.sparse.h"
+#include"MMSP.sparse.cpp"
+
+#include"MMSP.grid.h"
+#include"MMSP.grid.cpp"
